+++ .idea/workspace.xml
@@ -5,17 +5,37 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="1438e983-6e19-43a2-84fd-50f8c277f320" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/DVRS/CustomerAgent.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/DeliveryAgent1.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/DeliveryAgent2.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/DeliveryAgent3.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/DeliveryAgent4.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/MasterRoutingAgent.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/Parcel.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/ParcelList.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/DVRS/package-info.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.project" beforeDir="false" afterPath="$PROJECT_DIR$/.project" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/APDescription.txt" beforeDir="false" afterPath="$PROJECT_DIR$/APDescription.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/MTPs-Main-Container.txt" beforeDir="false" afterPath="$PROJECT_DIR$/MTPs-Main-Container.txt" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/CustomerAgent$ParcelTicker.class" beforeDir="false" afterPath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/CustomerAgent$ParcelTicker.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/CustomerAgent.class" beforeDir="false" afterPath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/CustomerAgent.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/MasterRoutingAgent$1.class" beforeDir="false" afterPath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/MasterRoutingAgent$1.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/MasterRoutingAgent.class" beforeDir="false" afterPath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/MasterRoutingAgent.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/Parcel.class" beforeDir="false" afterPath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/Parcel.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/allAgents/CustomerAgent.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/allAgents/CustomerAgent.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/allAgents/MasterRoutingAgent.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/allAgents/MasterRoutingAgent.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/CustomerAgent$ParcelTicker.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/CustomerAgent.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/DeliveryAgent$1.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/DeliveryAgent.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/DeliveryAgent1.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/DeliveryAgent2.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/DeliveryAgent3.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/MasterRoutingAgent$1.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/MasterRoutingAgent.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/OptimizationTechniques.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/classes/production/COS30018_Assignment1/allAgents/Parcel.class" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/CustomerAgent.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/DeliveryAgent.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/DeliveryAgent1.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/DeliveryAgent2.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/DeliveryAgent3.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/MasterRoutingAgent.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/OptimizationTechniques.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/allAgents/package-info.java" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -26,6 +46,7 @@
     <option name="RECENT_TEMPLATES">
       <list>
         <option value="Class" />
+        <option value="package-info" />
       </list>
     </option>
   </component>
@@ -82,7 +103,7 @@
     <configuration name="COS30018_Assignment1" type="Application" factoryName="Application">
       <option name="MAIN_CLASS_NAME" value="jade.Boot" />
       <module name="COS30018_Assignment1" />
-      <option name="PROGRAM_PARAMETERS" value="-gui -agents &quot;customerAgent:allAgents.CustomerAgent;masterroutingAgent:allAgents.MasterRoutingAgent;deliveryAgent:allAgents.DeliveryAgent;deliveryAgent1:allAgents.DeliveryAgent1;deliveryAgent2:allAgents.DeliveryAgent2;deliveryAgent3:allAgents.DeliveryAgent3&quot;" />
+      <option name="PROGRAM_PARAMETERS" value="-gui -agents &quot;customerAgent:DVRS.CustomerAgent;masterroutingAgent:DVRS.MasterRoutingAgent;deliveryAgent1:DVRS.DeliveryAgent1;deliveryAgent2:DVRS.DeliveryAgent2;deliveryAgent3:DVRS.DeliveryAgent3;deliveryAgent4:DVRS.DeliveryAgent4&quot;" />
       <method v="2">
         <option name="Make" enabled="true" />
       </method>
