+++ src/DVRS/DeliveryAgent2.java
@@ -1,0 +1,10 @@
+package DVRS;
+
+public class DeliveryAgent2 extends DeliveryAgent1
+{
+    protected void setup()
+    {
+        super.setup();
+        maxCapacity=-50;
+    }
+}
