+++ src/DVRS/DeliveryAgent3.java
@@ -1,0 +1,10 @@
+package DVRS;
+
+public class DeliveryAgent3 extends DeliveryAgent1
+{
+    protected void setup()
+    {
+        super.setup();
+        maxCapacity=200;
+    }
+}
