+++ src/DVRS/DeliveryAgent4.java
@@ -1,0 +1,10 @@
+package DVRS;
+
+public class DeliveryAgent4 extends DeliveryAgent1
+{
+    protected void setup()
+    {
+        super.setup();
+        maxCapacity=500;
+    }
+}
