+++ src/DVRS/Parcel.java
@@ -1,0 +1,15 @@
+package DVRS;
+
+public class Parcel {
+    public String name;
+    public int weight;
+    public int x;
+    public int y;
+
+    public Parcel(String name, int weight, int x, int y) {
+        this.name = name;
+        this.weight = weight;
+        this.x = x;
+        this.y=y;
+    }
+}
