+++ src/DVRS/package-info.java
@@ -1,8 +1,8 @@
 /**
- * 
+ *
  */
 /**
  * @author USER
  *
  */
-package allAgents;
+package DVRS;
